requirements.txt
bunch>=1.0.1 # bunch in vessel_segmentation prevents a python version >= 3.10
captum>=0.7.0
gdown
imgaug>=0.4.0
ipykernel>=6.15.2
ipython>=8.16.1
jupyter>=1.0.0
jupyter-black>=0.3.4
matplotlib>=3.6.2
matplotlib-inline>=0.1.6
munch>=4.0.0
mypy>=0.910
notebook_shim>=0.2.4
numpy>=1.24.0
opencv-python-headless>=4.9.0.80
openpyxl>=3.1.2
pandas>=1.4.2
pillow>=10.2.0
PyYAML>=6.0
scikit-image>=0.22.0
scikit-learn>=1.1.2
seaborn>=0.13.2
timm>=0.9.16
torch>=2.2.0
torchinfo>=1.8.0
torchvision>=0.15.2
tqdm>=4.66.2
scipy>=1.10.0
albumentations # no version to prevent conflict with imaug